+++ src/ChampionsChromo.Application/Schools/Commands/UpdateSchool/UpdateSchoolCommand.cs
@@ -14,6 +14,6 @@ public record UpdateSchoolCommand : IRequest<Result>
     public string State { get; set; } = string.Empty;
     public string Warning { get; set; } = string.Empty;
     public string BgWarningColor { get; set; } = string.Empty;
-    public int ShippingCost { get; set; }
+    public decimal ShippingCost { get; set; }
 
 }
